revise Lipschitz constant computation
